% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flickr_userdays}
\alias{flickr_userdays}
\title{Popularity of Flickr, in user-days}
\format{
A time series object with 156 observations.
}
\source{
Flickr. (2019). Retrieved October, 2019, fromhttps://flickr.com/
}
\usage{
flickr_userdays
}
\description{
A time series representing the popularity of flickr in the United States, as measured in user-days. Here, user-days count the number of unique users posting on flickr on a given day.
}
\keyword{datasets}
