% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_parameter_df.R
\name{generate_parameter_df}
\alias{generate_parameter_df}
\title{Generate a parameter dataframe}
\usage{
generate_parameter_df(parkdata, popularity_proxy, use_ref_series = FALSE, ...)
}
\arguments{
\item{parkdata}{A data frame of park visitation data in with columns as in the park_visitation dataset.}

\item{popularity_proxy}{A time series object representing the popularity of social media used for PUD.}

\item{use_ref_series}{A boolean specifying whether nps data is to be used for parameter estimates.}

\item{...}{Additional arguments to be passed onto the visitation_model function.}
}
\description{
For a given dataframe of parks and popularity proxy, generates a data frame of parameters.
}
