% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gtrends_popularity}
\alias{gtrends_popularity}
\title{A dataset containing google trends data for the keyword "flickr", accessed on 7/23/20.}
\format{
An object of class \code{ts} of length 156.
}
\usage{
gtrends_popularity
}
\description{
A dataset containing google trends data for the keyword "flickr", accessed on 7/23/20.
}
\keyword{datasets}
