% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{park_visitation}
\alias{park_visitation}
\title{National park visitation counts and associated Photo-User-Days data}
\format{
A data frame with 3276 rows and 3 variables.

\describe{
  \item{date}{date of monthly observation, in year-month-day format.}
  \item{park}{national park alpha code identifying a national park.}
  \item{pud}{Flickr photo-user-days (pud).  Here, photo-user-days count the number of unique users posting a photo on flickr on a given day from within the boundaries of a given national park. }
}
}
\source{
Flickr. (2019). Retrieved October, 2019, from https://flickr.com/
}
\usage{
park_visitation
}
\description{
A data frame storing monthtly flickr photo-user-days (pud) for 20 popular US national parks.
Here, photo-user-days count the number of unique users posting a photo on flickr on a given day from within the boundaries of a given national park.
}
\keyword{datasets}
