% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_sliding_window.R
\name{sliding_window_VisitorCounts}
\alias{sliding_window_VisitorCounts}
\title{Sliding Window Results}
\usage{
sliding_window_VisitorCounts(
  parkdata,
  popularity_proxy,
  model = c("pud_only", "nps_assisted", "nps_only"),
  lag_method = "cross-correlation",
  num_windows = 48,
  n_ahead = 12,
  suspected_periods = c(12, 6, 4, 3),
  proportion_of_variance_type = c("leave_out_first", "total"),
  max_proportion_of_variance = 0.995,
  log_ratio_cutoff = 0.2,
  window_length = "auto",
  num_trend_components = 2,
  trace = TRUE,
  ...
)
}
\arguments{
\item{parkdata}{A dataframe matching the format of \code{park_visitation} from the VisitorCounts package.}

\item{popularity_proxy}{A time series matching the format of \code{flickr_userdays} from the VisitorCounts package.}

\item{model}{A character string specifiying the option for which model to assess. One of "pud_only" (uses only the pud data), "nps_assisted" (uses nps data for parameter estimates), "nps_only" (forecasts using SSA using just the nps data)}

\item{num_windows}{The number of windows to use in the sliding window analysis.}

\item{n_ahead}{An integer describing the number of forecasts to make in each window.}

\item{suspected_periods}{A vector which stores the suspected periods in the descending order of importance. The default option is c(12,6,4,3), corresponding to 12, 6, 4, and 3 months.}

\item{proportion_of_variance_type}{A character string specifying the option for choosing the maximum number of eigenvalues based on the proportion of total variance explained. If "leave_out_first" is chosen, then the contribution made by the first eigenvector is ignored; otherwise, if "total" is chosen, then the contribution made by all the eigenvectors is considered.}

\item{max_proportion_of_variance}{A numeric specifying the proportion of total variance explained using the method specified in proportion_of_variance_type. The default option is 0.995.}

\item{log_ratio_cutoff}{A numeric specifying the threshold for the deviation between the estimated period and candidate periods in suspected_periods. THe default option is 0.2, which means that if the absolute log ratio between the estimated and candidate period is within 0.2 (approximately a 20 percent difference), then the estimated period is deemed equal to the candidate period.}

\item{window_length}{A character string or positive integer specifying the window length for the SSA estimation. If "auto" is chosen, then the algorithm automatically selects the window length by taking a multiple of 12 which does not exceed half the length of time_series. The default option is "auto".}

\item{num_trend_components}{A positive integer specifying the number of eigenvectors to be chosen for describing the trend in SSA. The default option is 2.}

\item{trace}{A boolean specifying whether to print out an index whenever the sliding window is completed for a given park.}

\item{...}{Additional arguments to be passed onto the visitation_model function.}
}
\value{
\item{park_MAE_diff_mat}{Mean Absolute Error for forecasted differences in sliding window}
\item{park_RMSE_diff_mat}{Root-Mean-Square Error for forecasted differences in sliding window}
}
\description{
Sliding window used for model evaluation in \emph{Modeling and Forecasting Percent Changes in National ParkVisitation Using Social Media}.
}
